get alpha working in sbre with vertex shaders enabled. some work on point sprite nonsense for sfx and lasers. code thrash due to vector3.h change
git-svn-id: https://pioneer.svn.sourceforge.net/svnroot/pioneer/trunk@392 e632f14b-6550-0410-b89e-a82653faca30

+++ src/CityOnPlanet.cpp
@@ -197,7 +197,7 @@ CityOnPlanet::CityOnPlanet(const Planet *planet, const SpaceStation *station, Ui
 		int a, b;
 		a = rand.Int32(-1000,1000);
 		b = rand.Int32(-1000,1000);
-		cityflavour[i].center = p + a*mx + b*mz;
+		cityflavour[i].center = p + (double)a*mx + (double)b*mz;
 		cityflavour[i].size = (double)rand.Int32(blist->minRadius, blist->maxRadius);
 	}
